refactor: navigationBar 구조 분리
